rate bug fix in home and find worker view

+++ iAssist/Models/BidViewModel.cs
@@ -24,6 +24,6 @@ namespace iAssist.Models
         public int? workerid { get; set; }
         public string Username { get; set; }
         public int? bookstatus { get; set; }
-        public int? Rate { get; set; }
+        public double? Rate { get; set; }
     }
 }
